Rework printing
The types method for printing is changed from 'format' to
'print_object'.  This mimics Common Lisp's 'print-object'.

A method 'stringify' is also added.  It returns a string like
'<#{type} 0x{hexaddr}>'.  The T type has a generic implementation,
and other types are allowed to specialize it, but should retain the
generic format (and example is the NUMBER type, which also includes
the value in this string).  This method allows other usages, such as
debugging and introspection.

The T type also gets a generic method 'print_object', which simply
prints the result from 'stringify' on the same object.  Other types
should override it for better prints.